changed to run register on appstart and not as separate script (workaround for avoiding env issues)
