+++ .vscode/settings.json
@@ -1,10 +1,12 @@
 {
     "cSpell.words": [
         "appuser",
+        "GITHUB",
         "kemur",
         "mimimi",
         "MIMIMI",
         "pawnobi",
+        "sibiraj",
         "Wael"
     ],
     "editor.codeActionsOnSave": {
